+++ src/modules/mavlink/CMakeLists.txt
@@ -58,8 +58,8 @@ px4_add_module(
 		mavlink_shell.cpp
 		mavlink_simple_analyzer.cpp
 		mavlink_stream.cpp
-		mavlink_ulog.cpp
 		mavlink_timesync.cpp
+		mavlink_ulog.cpp
 		tune_publisher.cpp
 	MODULE_CONFIG
 		module.yaml
